Revert "Gcb 45"
