--- DataLayer/Interfaces/INormsRepository.cs
@@ -1,9 +1,0 @@
-﻿using bumbo.Models;
-
-namespace DataLayer.Interfaces
-{
-    public interface INormsRepository
-    {
-        List<Norm> GetNorms();
-    }
-}
