--- DataLayer/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-{
-  "profiles": {
-    "DataLayer": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "https://localhost:51606;http://localhost:51607"
-    }
-  }
-}
