+++ bumbo.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35222.181
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1706.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bumbo", "bumbo\bumbo.csproj", "{00F164C0-1EB2-449C-86F0-E519AFDD035C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataLayer", "DataLayer\DataLayer.csproj", "{99161EFE-3C8F-40DE-80DC-DBCCB62AD5F9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "bumbo", "bumbo\bumbo.csproj", "{00F164C0-1EB2-449C-86F0-E519AFDD035C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +15,6 @@ Global
 		{00F164C0-1EB2-449C-86F0-E519AFDD035C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{00F164C0-1EB2-449C-86F0-E519AFDD035C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{00F164C0-1EB2-449C-86F0-E519AFDD035C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{99161EFE-3C8F-40DE-80DC-DBCCB62AD5F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{99161EFE-3C8F-40DE-80DC-DBCCB62AD5F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{99161EFE-3C8F-40DE-80DC-DBCCB62AD5F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{99161EFE-3C8F-40DE-80DC-DBCCB62AD5F9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
