--- bumbo/Controllers/AccountController.cs
@@ -1,56 +1,0 @@
-﻿using bumbo.Models;
-using bumbo.ViewModels;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-
-public class AccountController : Controller
-{
-    private readonly SignInManager<Employee> _signInManager;
-    private readonly UserManager<Employee> _userManager;
-
-    public AccountController(SignInManager<Employee> signInManager, UserManager<Employee> userManager)
-    {
-        _signInManager = signInManager;
-        _userManager = userManager;
-    }
-
-    [HttpGet]
-    public IActionResult Login()
-    {
-        return View();
-    }
-
-    [HttpPost]
-    public async Task<IActionResult> Login(LoginViewModel model)
-    {
-        if (ModelState.IsValid)
-        {
-            var user = await _userManager.FindByEmailAsync(model.Email);
-            if (user == null)
-            {
-                ModelState.AddModelError("", "Dit e-mailadres bestaat niet.");
-            }
-            else
-            {
-                var result = await _signInManager.PasswordSignInAsync(user, model.Password, model.RememberMe, lockoutOnFailure: false);
-
-                if (result.Succeeded)
-                {
-                    return RedirectToAction("Index", "Home");
-                }
-                else
-                {
-                    ModelState.AddModelError("", "Wachtwoord is onjuist.");
-                }
-            }
-        }
-
-        return View(model);
-    }
-
-    public async Task<IActionResult> Logout()
-    {
-        await _signInManager.SignOutAsync();
-        return RedirectToAction("Index", "Home");
-    }
-}
