--- bumbo/ViewModels/LoginViewModel.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace bumbo.ViewModels
-{
-    public class LoginViewModel
-    {
-        [Required(ErrorMessage = "Email is required.")]
-        [EmailAddress(ErrorMessage = "Invalid Email Address.")]
-        public string Email { get; set; }
-
-        [Required(ErrorMessage = "Password is required.")]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-
-        [Display(Name = "Remember me?")]
-        public bool RememberMe { get; set; }
-    }
-}
