--- bumbo/ViewModels/NavBarViewModel.cs
@@ -1,13 +1,0 @@
-﻿namespace bumbo.ViewModels
-{
-    public class NavBarViewModel
-    {
-        public string FirstName { get; set; }
-        public string Email { get; set; }
-        public bool IsAuthenticated { get; set; }
-        public bool IsSystemManager { get; set; }
-        public bool IsBranchManager { get; set; }
-
-    }
-
-}
