--- bumbo/ViewModels/Norms/ReadNormViewModel.cs
@@ -1,12 +1,0 @@
-﻿namespace bumbo.Models.ViewModels.Norms
-{
-    public class ReadNormViewModel
-    {
-        public int Week { get; set; }
-        public int UnloadColis { get; set; }
-        public int FillShelves { get; set; }
-        public int Cashier { get; set; }
-        public int Fresh { get; set; }
-        public int Fronting { get; set; }
-    }
-}
