--- bumbo/ViewModels/Norms/UpdateNormViewModel.cs
@@ -1,12 +1,0 @@
-﻿namespace bumbo.Models.ViewModels.Norms
-{
-    public class UpdateNormViewModel
-    {
-        public int FirstNormId { get; set; }
-        public int UnloadColis { get; set; }
-        public int FillShelves { get; set; }
-        public int Cashier { get; set; }
-        public int Fresh { get; set; }
-        public int Fronting { get; set; }
-    }
-}
