+++ bumbo/Views/Home/Index.cshtml
@@ -10,4 +10,3 @@
     <default-button button-type="secondary"> </default-button>
 </div>
 
-
