+++ bumbo/Views/Norms/Index.cshtml
@@ -1,11 +1,6 @@
-﻿@model IEnumerable<bumbo.Models.Norm>
-
-@{
+﻿@{
     ViewData["Title"] = "Normeringen";
 }
 
-<div>
-    @Html.Raw(ViewBag.HtmlTable)
-</div>
-
-<a asp-controller="Norms" asp-action="Create" asp-route-lastWeek="false" class="btn btn-primary">Nieuwe normering</a>
+<h1>Normeringen Page</h1>
+<p>This is the Normeringen page.</p>
