+++ bumbo/Views/Shared/Components/NavBarViewComponent.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using bumbo.Models;
-using bumbo.ViewModels;
 
 namespace bumbo.ViewComponents
 {
@@ -42,3 +41,12 @@ namespace bumbo.ViewComponents
 }
 
 
+public class NavBarViewModel
+{
+        public string FirstName { get; set; }
+        public string Email { get; set; }
+        public bool IsAuthenticated { get; set; }
+        public bool IsSystemManager { get; set; }
+        public bool IsBranchManager { get; set; }
+
+}
