+++ bumbo/Views/_ViewImports.cshtml
@@ -1,5 +1,6 @@
 ﻿@using bumbo
 @using bumbo.Models
-@using bumbo.Models.ViewModels.Norms
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, bumbo
+
+
