+++ bumbo/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "bumbo": "Server=localhost;Database=Bumbo;Trusted_Connection=True;Encrypt=False;"
+    "anthony": "Server=ANTHONY;Database=Bumbo;Trusted_Connection=True;Encrypt=False;"
   }
 }
