+++ bumbo/wwwroot/css/site.css
@@ -2,12 +2,6 @@
     font-size: 14px;
 }
 
-body {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
 @media (min-width: 768px) {
     html {
         font-size: 16px;
@@ -18,6 +12,14 @@ body {
     box-shadow: 0 0 0 0.1rem white, 0 0 0 0.25rem #258cfb;
 }
 
+html {
+    position: relative;
+    min-height: 100%;
+}
+
+body {
+    margin-bottom: 60px;
+}
 
 .primary_bg_color {
     background-color: #EFB716;
